fix(nickname): long -> int 타입 불일치 수정

+++ src/main/java/com/dreamypatisiel/devdevdev/domain/entity/Member.java
@@ -196,7 +196,7 @@ public class Member extends BasicTime {
         this.nicknameUpdatedAt = now;
     }
 
-    public boolean canChangeNickname(long restrictionHours) {
+    public boolean canChangeNickname(int restrictionHours) {
         return nicknameUpdatedAt == null
                 || ChronoUnit.HOURS.between(nicknameUpdatedAt, LocalDateTime.now()) >= restrictionHours;
     }
